requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy[asyncio]==2.0.30
asyncpg==0.29.0
temporalio==1.3.0
pydantic==2.6.4
jinja2==3.1.3
GitPython==3.1.43
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
